refactor: re-merge into one module, refactor OOP-ily
